refactor : 리뷰 피드백 반영 및 N+1 문제 최적화
- OrderRepository에 @EntityGraph 적용하여 N+1 성능 문제 해결
- 금액 데이터의 정확성을 위해 price 필드 타입을 Long에서 BigDecimal로 변경
- OrderMapper 내 연관 엔티티(Product, Option) null 체크 로직 추가
- Controller 내 페이지 사이즈 검증 로직 가독성 개선 (List.of 사용)
- DTO 및 Mapper 간 데이터 타입 불일치 해결

Close #40

+++ src/main/java/com/example/unbox_be/domain/order/dto/OrderResponseDto.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -16,7 +17,7 @@ import java.util.UUID;
 public class OrderResponseDto {
     // 1. 주문 기본 정보
     private UUID orderId;
-    private Long price;
+    private BigDecimal price;
     private OrderStatus status;
     private LocalDateTime createdAt;
